+++ components/cli/src/domain/use-cases/create-clean-tenant.ts
@@ -9,11 +9,14 @@ import type { FlySystem } from '../contracts/fly-system';
 import type { Operations } from '@crystallize/schema/mass-operation';
 import type { Logger } from '../contracts/logger';
 import { getMassOperationBulkTask } from '../../command/mass-operation/run';
+import type { InstallBoilerplateStore } from '../../ui/journeys/install-boilerplate/create-store';
 
 type Deps = {
     createCrystallizeClient: typeof createClient;
     runMassOperation: ReturnType<typeof createRunMassOperationHandler>;
     credentialsRetriever: CredentialRetriever;
+    crystallizeEnvironment: 'staging' | 'production';
+    installBoilerplateCommandStore: InstallBoilerplateStore;
     flySystem: FlySystem;
     logger: Logger;
 };
@@ -29,13 +32,27 @@ export type CreateCleanTenantHandlerDefinition = CommandHandlerDefinition<
     Awaited<ReturnType<typeof handler>>
 >;
 
+const sleep = (second: number) => new Promise((resolve) => setTimeout(resolve, second * 1000));
+
 const handler = async (
     envelope: Envelope<Command>,
-    { createCrystallizeClient, credentialsRetriever, runMassOperation, flySystem, logger }: Deps,
+    {
+        createCrystallizeClient,
+        credentialsRetriever,
+        runMassOperation,
+        flySystem,
+        logger,
+        crystallizeEnvironment,
+        installBoilerplateCommandStore,
+    }: Deps,
 ): Promise<{
     id: string;
     identifier: string;
 }> => {
+    const { storage, atoms } = installBoilerplateCommandStore;
+    const addTraceLog = (log: string) => storage.set(atoms.addTraceLogAtom, log);
+    const addTraceError = (log: string) => storage.set(atoms.addTraceErrorAtom, log);
+
     const { tenant, credentials } = envelope.message;
     const finalCredentials = credentials || (await credentialsRetriever.getCredentials());
     const client = createCrystallizeClient({
@@ -67,6 +84,7 @@ const handler = async (
         },
     );
     const { id, identifier } = createResult.tenant.create;
+    addTraceLog(`Tenant created with id: ${id}.`);
     const shapeIdentifiers = ['default-product', 'default-folder', 'default-document'];
     const mutation = {
         shape: shapeIdentifiers.reduce((memo: Record<string, any>, shapeIdentifier: string) => {
@@ -85,6 +103,7 @@ const handler = async (
     };
     const query = jsonToGraphQLQuery({ mutation });
     await client.pimApi(query);
+    addTraceLog(`Shape cleaned.`);
 
     // if we have a folder, we check that folder for .crystallize folder and convention
     const { folder } = envelope.message;
@@ -115,18 +134,23 @@ const handler = async (
             },
         } as Envelope<RunMassOperationCommand>);
         logger.debug('Mass operation task created', startedTask);
+        addTraceLog(`Mass operation task created: ${startedTask?.id}`);
+        await sleep(10); // we have an easy 10 sec sleep here to let the task start
         while (startedTask?.status !== 'complete') {
-            await new Promise((resolve) => setTimeout(resolve, 1000));
             const get = await cClient.nextPimApi(getMassOperationBulkTask, { id: startedTask?.id });
             if (get.bulkTask.error) {
                 throw new Error(get.data.bulkTask.error);
             }
             startedTask = get.bulkTask;
+            await sleep(3); // then we check every 3 seconds
         }
     } catch (e) {
+        addTraceError(`Failed to run mass operation.`);
         logger.error('Failed to run mass operation', e);
     }
+    addTraceLog(`Mass operation completed.`);
 
+    // now the extra mutations
     try {
         const results = await cClient.pimApi(`#graphql
             query { 
@@ -156,15 +180,32 @@ const handler = async (
         }[];
 
         for (const { mutation, sets } of Object.values(extraMutations)) {
-            await Promise.all([
+            await Promise.all(
                 sets.map(async (set) => {
                     await cClient.pimApi(mutation, set);
                 }),
-            ]);
+            );
         }
     } catch (e) {
+        addTraceError(`Failed to run extra mutations.`);
         logger.error('Failed to run extra mutations', e);
     }
+
+    // now the index
+    await cClient.nextPimApi(`mutation { igniteTenant }`);
+    // check the 404
+
+    const discoHost = crystallizeEnvironment === 'staging' ? 'api-dev.crystallize.digital' : 'api.crystallize.com';
+    let discoApiPingResponseCode = 404;
+
+    await sleep(15); // easy 15 sec sleep to let the index finish
+    do {
+        const discoApiPingResponse = await fetch(`https://${discoHost}/${identifier}/discovery`);
+        discoApiPingResponseCode = discoApiPingResponse.status;
+        sleep(5); // then every 5 seconds
+    } while (discoApiPingResponseCode === 404);
+
+    addTraceLog(`Tenant ignited in Discovery API.`);
     return {
         id,
         identifier,